added new tests

+++ src/main/amazed/Test.java
@@ -5,7 +5,7 @@ import amazed.maze.Amazed;
 public class Test {
 
     public static void main(String[] args) {
-        double timeDef = 0; 
+        long timeDef = 0; 
         double start = System.currentTimeMillis(); 
         System.out.println(start);
         long numberOfRuns = 0;
@@ -13,13 +13,14 @@ public class Test {
         
         Amazed amazed = new Amazed("maps/medium.map", false, 3, -1);
 
+        //runs alot of test
         for(int i = 0; i <= maxRuns; i++){
             
             amazed.solve();
             numberOfRuns = i; 
         }
 
-        timeDef = System.currentTimeMillis() - start; 
+        timeDef = System.currentTimeMillis() - (long)start; 
        
         double result = timeDef / maxRuns;
